fix warning on solaris - isspace is a macro that expects to be given an int
git-svn-id: svn://svn.code.sf.net/p/tivodecode/code/tivodecode/trunk@218 3cae7af1-9334-450b-b769-427e0023cdf8

+++ tivodecode.c
@@ -186,7 +186,7 @@ int main(int argc, char *argv[])
             if (fread(mak, 1, 11, mak_file) >= 10)
             {
                 int i;
-                for (i = 11; i >= 10 && (mak[i] == '\0' || isspace(mak[i])); --i)
+                for (i = 11; i >= 10 && (mak[i] == '\0' || isspace((int)mak[i])); --i)
                 {
                     mak[i] = '\0';
                 }
